backend/requirements.txt
addict==2.4.0
annotated-types==0.7.0
asgiref==3.11.0
attrs==25.4.0
babel==2.17.0
blis==1.3.3
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
csvw==3.7.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.13
Django==5.2.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dlinfo==2.0.0
docopt==0.6.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
espeakng-loader==0.2.4
filelock==3.20.3
fsspec==2026.1.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
isodate==0.7.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kokoro==0.9.4
language-tags==1.2.0
loguru==0.7.3
MarkupSafe==3.0.3
misaki==0.9.4
mpmath==1.3.0
murmurhash==1.0.15
networkx==3.6.1
num2words==0.5.14
numpy==2.4.1
packaging==25.0
phonemizer-fork==3.3.2
pillow==12.1.0
preshed==3.0.12
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
rdflib==7.5.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rfc3986==1.5.0
rpds-py==0.30.0
safetensors==0.7.0
segments==2.3.0
six==1.17.0
smart_open==7.5.0
soundfile==0.13.1
spacy==3.8.11
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.5
srsly==2.5.2
sympy==1.14.0
termcolor==3.3.0
thinc==8.3.10
tokenizers==0.22.2
torch==2.9.1
tqdm==4.67.1
transformers==4.57.6
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
wasabi==1.1.3
weasel==0.4.3
wrapt==2.0.1
